+++ recognition/include/pcl/recognition/surface_normal_modality.h
@@ -672,7 +672,7 @@ pcl::SurfaceNormalModality<PointInT>::computeAndQuantizeSurfaceNormals ()
         surface_normals_.points[index].normal_z = normal_z;
         surface_normals_.points[index].curvature = bad_point;
 
-        float angle = 11.25f + atan2 (normal_y, normal_x)*180.0f/3.14f;
+        float angle = 11.25f + std::atan2 (normal_y, normal_x)*180.0f/3.14f;
 
         if (angle < 0.0f) angle += 360.0f;
         if (angle >= 360.0f) angle -= 360.0f;
@@ -892,7 +892,7 @@ pcl::SurfaceNormalModality<PointInT>::computeAndQuantizeSurfaceNormals2 ()
           l_ny *= l_norminv;
           l_nz *= l_norminv;
 
-          float angle = 22.5f + atan2f (l_ny, l_nx) * 180.0f / 3.14f;
+          float angle = 22.5f + std::atan2 (l_ny, l_nx) * 180.0f / 3.14f;
 
           if (angle < 0.0f) angle += 360.0f;
           if (angle >= 360.0f) angle -= 360.0f;
@@ -1378,7 +1378,7 @@ pcl::SurfaceNormalModality<PointInT>::quantizeSurfaceNormals ()
       //quantized_surface_normals_.data[row_index*width+col_index] =
       //  normal_lookup_(normal_x, normal_y, normal_z);
 
-      float angle = 11.25f + atan2f (normal_y, normal_x)*180.0f/3.14f;
+      float angle = 11.25f + std::atan2 (normal_y, normal_x)*180.0f/3.14f;
 
       if (angle < 0.0f) angle += 360.0f;
       if (angle >= 360.0f) angle -= 360.0f;
